Move more common utils to internal-testutils
