+++ gradle/libs.versions.toml
@@ -51,10 +51,12 @@ androidx-lifecycle-viewmodel-compose = "androidx.lifecycle:lifecycle-viewmodel-c
 
 napier = "io.github.aakira:napier:1.4.1"
 
+androidx-test-core = { module = "androidx.test:core-ktx", version.ref = "androidxtest" }
 androidx-test-runner = { module = "androidx.test:runner", version.ref = "androidxtest" }
 androidx-test-rules = { module = "androidx.test:rules", version.ref = "androidxtest" }
 androidx-test-orchestrator = { module = "androidx.test:orchestrator", version.ref = "androidxtest" }
 androidx-test-uiAutomator = "androidx.test.uiautomator:uiautomator:2.2.0"
+
 junit = "junit:junit:4.13"
 truth = "com.google.truth:truth:1.1.2"
 robolectric = "org.robolectric:robolectric:4.5.1"
